--- a/pkg/engine/planner/logical/planner_test.go
+++ b/pkg/engine/planner/logical/planner_test.go
@@ -83,7 +83,7 @@ func TestConvertAST_Success(t *testing.T) {
 		statement: `{cluster="prod", namespace=~"loki-.*"} | foo="bar" or bar="baz" |= "metric.go" |= "foo" or "bar" !~ "(a|b|c)" `,
 		start:     3600,
 		end:       7200,
-		direction: logproto.FORWARD,
+		direction: logproto.BACKWARD, // ASC is not supported
 		limit:     1000,
 	}
 	logicalPlan, err := BuildPlan(q)
@@ -94,7 +94,7 @@ func TestConvertAST_Success(t *testing.T) {
 %2 = MATCH_RE label.namespace "loki-.*"
 %3 = AND %1 %2
 %4 = MAKETABLE [selector=%3, shard=0_of_1]
-%5 = SORT %4 [column=builtin.timestamp, asc=true, nulls_first=false]
+%5 = SORT %4 [column=builtin.timestamp, asc=false, nulls_first=false]
 %6 = GTE builtin.timestamp 1970-01-01T01:00:00Z
 %7 = SELECT %5 [predicate=%6]
 %8 = LT builtin.timestamp 1970-01-01T02:00:00Z
@@ -243,7 +243,7 @@ func TestCanExecuteQuery(t *testing.T) {
 				statement: tt.statement,
 				start:     1000,
 				end:       2000,
-				direction: logproto.FORWARD,
+				direction: logproto.BACKWARD,
 				limit:     1000,
 			}