### What this PR does / why we need it

Sorting sections by `timestamp DESC` first and `streamID ASC` second allows for more efficient reading (by section) for BACKWARD log queries.

### Special notes for your reviewer

- Supersedes work in progress PR https://github.com/grafana/loki/pull/18337
- This change breaks FORWARD log queries. Which is ok for the time being. We will add support later again when columnar reading of the dataobj is introduced.
- This change renders existing dataobj in storage incompatible.

### Benchmarks

<details>

```console
$ benchstat logql_bench_main.txt logql_bench_sortbytimestamp.txt
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/v3/pkg/logql/bench
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                                                                                                                                                              │ logql_bench_main.txt │   logql_bench_sortbytimestamp.txt    │
                                                                                                                                                              │        sec/op        │    sec/op     vs base                │
LogQL/query={region="ap-southeast-1",_env="dev"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                           10301.8m ± 45%   776.6m ± 12%  -92.46% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                     7.865 ± 58%    1.577 ± 27%  -79.95% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                      8.327 ± 29%    1.709 ±  3%  -79.48% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                               11.894 ± 30%    1.548 ± 32%  -86.98% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                        15.766 ± 22%    1.377 ±  8%  -91.27% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                               13.888 ± 39%    1.234 ± 41%  -91.11% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                12.109 ± 29%    1.200 ±  2%  -90.09% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                          18.508 ± 44%    1.649 ± 29%  -91.09% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                          7.060 ± 13%    1.769 ± 27%  -74.94% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                 3.689 ± 33%    3.576 ± 13%        ~ (p=0.971 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                  4.760 ± 26%    2.565 ±  1%  -46.12% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8            4.713 ± 40%    3.538 ± 14%  -24.93% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                      13.157 ± 27%    1.208 ±  6%  -90.82% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                              8.846 ± 25%    1.577 ± 32%  -82.17% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                               9.142 ± 27%    1.008 ±  7%  -88.98% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                        10.877 ± 27%    1.409 ± 29%  -87.04% (p=0.000 n=10)
geomean                                                                                                                                                                  9.206          1.591        -82.71%

                                                                                                                                                              │ logql_bench_main.txt │       logql_bench_sortbytimestamp.txt       │
                                                                                                                                                              │  kilobytesProcessed  │ kilobytesProcessed  vs base                 │
LogQL/query={region="ap-southeast-1",_env="dev"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                              2.502k ± 0%          3.990k ± 0%   +59.47% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                     2.502k ± 0%          3.990k ± 0%   +59.47% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                      2.502k ± 0%          4.005k ± 0%   +60.07% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                2.502k ± 0%          4.608k ± 0%   +84.17% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                         2.509k ± 0%          5.699k ± 0%  +127.14% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                2.509k ± 0%          5.699k ± 0%  +127.14% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                 2.509k ± 0%          5.699k ± 0%  +127.14% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                           2.509k ± 0%          5.234k ± 0%  +108.61% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                          2.501k ± 0%          5.140k ± 0%  +105.52% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                 2.501k ± 0%         11.683k ± 0%  +367.13% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                  2.501k ± 0%         11.936k ± 0%  +377.25% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8            2.501k ± 0%         12.089k ± 0%  +383.37% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                       2.495k ± 0%          3.368k ± 0%   +34.99% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                              2.495k ± 0%          4.160k ± 0%   +66.73% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                               2.495k ± 0%          4.160k ± 0%   +66.73% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                         2.495k ± 0%          3.864k ± 0%   +54.87% (p=0.000 n=10)
geomean                                                                                                                                                                  2.502k               5.421k       +116.67%

                                                                                                                                                              │ logql_bench_main.txt │     logql_bench_sortbytimestamp.txt      │
                                                                                                                                                              │    linesProcessed    │ linesProcessed  vs base                  │
LogQL/query={region="ap-southeast-1",_env="dev"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                              22.76k ± 0%      68.83k ± 0%   +202.45% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                     22.76k ± 0%      68.83k ± 0%   +202.45% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                      22.76k ± 0%      69.34k ± 0%   +204.70% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                22.76k ± 0%      97.50k ± 0%   +328.45% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                         22.76k ± 0%      68.83k ± 0%   +202.45% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                22.76k ± 0%      68.83k ± 0%   +202.45% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                 22.76k ± 0%      68.83k ± 0%   +202.45% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                           22.76k ± 0%      68.83k ± 0%   +202.45% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                          22.76k ± 0%     322.48k ± 0%  +1317.11% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                 22.76k ± 0%    1074.61k ± 0%  +4622.29% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                  22.76k ± 0%    1103.79k ± 0%  +4750.54% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8            22.76k ± 0%    1142.19k ± 0%  +4919.29% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                       22.76k ± 0%      68.83k ± 0%   +202.45% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                              22.76k ± 0%     112.35k ± 0%   +393.70% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                               22.76k ± 0%     112.35k ± 0%   +393.70% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                         22.76k ± 0%     112.86k ± 0%   +395.95% (p=0.000 n=10)
geomean                                                                                                                                                                  22.76k           143.1k        +528.67%

                                                                                                                                                              │ logql_bench_main.txt │     logql_bench_sortbytimestamp.txt      │
                                                                                                                                                              │   postFilterLines    │ postFilterLines  vs base                 │
LogQL/query={region="ap-southeast-1",_env="dev"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                             0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                    0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                     0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1",_env="dev"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                               0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                        0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                               0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                          0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                         0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                 0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8           0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="kubernetes"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                      0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="kubernetes"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                             0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="kubernetes"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                              0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="kubernetes"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                        0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                                                                                                                            ²                    +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

</details>